+++ stock-service/src/main/java/hanghae/stockservice/external/controller/dto/StockDto.java
@@ -4,10 +4,10 @@ public record StockDto(
 
         Long productId,
 
-        Long quantity
+        Integer quantity
 ) {
 
-    public static StockDto of(Long productId, Long quantity) {
+    public static StockDto of(Long productId, Integer quantity) {
         return new StockDto(productId, quantity);
     }
 }
