+++ stock-service/src/main/java/hanghae/stockservice/external/controller/dto/request/QuantityCheckRequest.java
@@ -2,6 +2,6 @@ package hanghae.stockservice.external.controller.dto.request;
 
 public record QuantityCheckRequest(
         Long productId,
-        Long orderQuantity
+        Integer orderQuantity
 ) {
 }
